# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: ghidra-rest
  version: "v1"
  summary: REST API in front of Ghidra's headless analyzer.
  description: |
    Upload a binary, poll the job, read back functions, decompiled C, strings,
    symbols, imports, exports, types, cross references and raw bytes as JSON.

    One job is one `analyzeHeadless` run that exports a complete artifact set
    and exits. Result endpoints only read those files, so they answer `409`
    until the job reaches `done`.

    Addresses accept `401000`, `0x401000`, `00401000` and `RAM:00401000`; they
    normalise to the same key (lower case, no `0x`, no leading zeros, address
    space prefix kept).
  license:
    name: See repository
...
servers:
  - url: http://127.0.0.1:8080
    description: Local container or binary

# Both schemes are the same shared token. Applied globally; it is enforced only
# when GHIDRAREST_API_TOKEN is set, and /healthz opts out entirely.
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /v1/capabilities:
    get:
      tags: [service]
      operationId: capabilities
      summary: Live limits, features and endpoint list
      description: For clients that would rather discover caps than hard code them.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Capabilities" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/jobs:
    get:
      tags: [jobs]
      operationId: listJobs
      summary: List jobs, newest first
      parameters:
        - name: status
          in: query
          description: Filter by lifecycle state.
          schema: { $ref: "#/components/schemas/JobStatus" }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer, description: Matching jobs before paging. }
                  count: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  jobs:
                    type: array
                    items: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Unauthorized" }

    post:
      tags: [jobs]
      operationId: submitJob
      summary: Submit a binary (multipart)
      description: |
        The binary goes in a part named `file`; the options are ordinary form
        fields. Identical bytes are deduplicated by sha256 unless `force` is
        set, which answers `200` with the existing job instead of burning
        another Ghidra run. Jobs that ended `failed` or `canceled` are not
        reused.

        Zero or absent numeric options mean "use the server default"; the
        returned job records the resolved values, so it shows what actually ran.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The binary. Analysed, never executed.
                name:
                  type: string
                  description: |
                    Stored name. Defaults to the upload's filename. Reduced to
                    a sanitised basename, 128 chars max.
                decompile:
                  type: boolean
                  description: Decompile functions during the run. Default `GHIDRAREST_DECOMPILE`.
                decompile_max_funcs:
                  type: integer
                  description: Stop decompiling after this many functions. Analysis still completes.
                decompile_timeout_sec:
                  type: integer
                  description: Per-function decompiler timeout.
                analysis_timeout_sec:
                  type: integer
                  description: |
                    Wall-clock budget for the whole job. Ghidra's own
                    `-analysisTimeoutPerFile` is set to 3/4 of it so the
                    exporter still gets to run.
                processor:
                  type: string
                  description: Force a language id.
                  examples: ["x86:LE:64:default"]
                compiler_spec:
                  type: string
                  examples: ["gcc"]
                loader:
                  type: string
                  examples: ["BinaryLoader"]
                force:
                  type: boolean
                  description: Analyse even if these exact bytes were analysed before.
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        "202": { $ref: "#/components/responses/JobAccepted" }
        "200": { $ref: "#/components/responses/JobDeduplicated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "503": { $ref: "#/components/responses/QueueFull" }
  /v1/jobs/base64:
    post:
      tags: [jobs]
      operationId: submitJobBase64
      summary: Submit a binary (JSON)
      description: |
        Same semantics as the multipart endpoint, for clients that would
        rather not build a multipart body. The body limit is
        `max_upload_bytes * 4/3 + 1 MiB`, since base64 costs a third more than
        the payload it carries.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data_base64]
              properties:
                data_base64:
                  type: string
                  format: byte
                  description: Standard base64. `data` is accepted as an alias.
                data:
                  type: string
                  format: byte
                  description: Alias for `data_base64`.
                filename: { type: string }
                force: { type: boolean }
                decompile:
                  type: [boolean, "null"]
                  description: Null or absent means the server default.
                decompile_max_funcs: { type: integer, description: 0 means the server default. }
                decompile_timeout_sec: { type: integer, description: 0 means the server default. }
                analysis_timeout_sec: { type: integer, description: 0 means the server default. }
                processor: { type: string }
                compiler_spec: { type: string }
                loader: { type: string }
      responses:
        "202": { $ref: "#/components/responses/JobAccepted" }
        "200": { $ref: "#/components/responses/JobDeduplicated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "503": { $ref: "#/components/responses/QueueFull" }
  /v1/jobs/{id}:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [jobs]
      operationId: getJob
      summary: One job record
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [jobs]
      operationId: deleteJob
      summary: Delete a job and its artifacts
      description: Removes the input, the log and every artifact. Cancel a live job first.
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Still queued or running
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/jobs/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/JobId"
    post:
      tags: [jobs]
      operationId: cancelJob
      summary: Cancel a queued or running job
      description: |
        A queued job is cancelled outright. A running one has its whole
        `analyzeHeadless` process group killed -- the launcher shell and the
        JVM it exec'd -- so no orphan keeps writing into a dead job.
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Already finished
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/jobs/{id}/log:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [jobs]
      operationId: getJobLog
      summary: The analyzeHeadless transcript
      description: |
        Starts with the exact argv that was run. Available while the job is
        still running, which is what you want on a long analysis.
      parameters:
        - name: tail
          in: query
          description: Return only the last N bytes.
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No such job, or no log written yet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/jobs/{id}/input:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [jobs]
      operationId: getJobInput
      summary: The submitted bytes back
      responses:
        "200":
          description: OK
          headers:
            Content-Disposition:
              schema: { type: string }
              description: Attachment, with the stored filename.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/jobs/{id}/export:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [jobs]
      operationId: exportJob
      summary: Every artifact as one zip
      description: |
        The artifact tree plus `meta.json` and `headless.log`, deflated and
        streamed rather than staged on disk.
      responses:
        "200":
          description: OK
          content:
            application/zip:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/summary:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [results]
      operationId: getSummary
      summary: Program summary
      description: |
        `summary.json` is written last by the export script, so its presence
        is what marks an artifact set complete.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Summary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/functions:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/Query"
    get:
      tags: [results]
      operationId: listFunctions
      summary: Functions, paged
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Function" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ArtifactMissing" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/function/{addr}:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Addr"
    get:
      tags: [results]
      operationId: getFunction
      summary: One function record
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Function" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No function starts at that address
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/function/{addr}/decompile:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Addr"
    get:
      tags: [results]
      operationId: decompileFunction
      summary: Decompiled C for one function
      parameters:
        - name: format
          in: query
          description: "`c` returns the listing alone as `text/x-c`."
          schema: { type: string, enum: [c] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Decompiled" }
            text/x-c:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: |
            No listing. The message distinguishes the causes: the job was
            submitted with `decompile=false`, the function is past
            `decompile_max_funcs`, it is external, or it is not a function.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/decompiled:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/Query"
    get:
      tags: [results]
      operationId: listDecompiled
      summary: Index of what was decompiled
      description: Cheaper than walking every function to find out.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/DecompiledIndexEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ArtifactMissing" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/disasm:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/Query"
    get:
      tags: [results]
      operationId: listDisasm
      summary: Instruction count per function
      description: Index over the disassembly artifacts, paged like the other lists.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/DisasmIndexEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ArtifactMissing" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/disasm/{addr}:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Addr"
    get:
      tags: [results]
      operationId: getDisasm
      summary: Instruction listing for one function
      description: >-
        Ghidra's listing for the function body, exported in the same pass as
        every other artifact. `?format=text` renders one instruction per line
        as text/plain instead of JSON.
      parameters:
        - name: format
          in: query
          required: false
          description: "`text` for a plain-text listing; omit for JSON."
          schema: { type: string, enum: [text] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DisasmListing" }
            text/plain:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: External, a data address, or a job analysed before this artifact existed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/xrefs/{addr}:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Addr"
    get:
      tags: [results]
      operationId: getXrefs
      summary: References to and from an address
      description: |
        Always `200` for a finished job. `indexed: false` with empty arrays
        means the address is not in the xref index, which distinguishes
        "nothing references this" from a typo'd address.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/XrefsResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ArtifactMissing" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/strings:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/Query"
    get:
      tags: [results]
      operationId: listStrings
      summary: Defined strings, paged
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/StringEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ArtifactMissing" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/symbols:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/Query"
    get:
      tags: [results]
      operationId: listSymbols
      summary: Symbol table, paged
      description: Dynamic symbols excluded.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Symbol" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ArtifactMissing" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/imports:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/Query"
    get:
      tags: [results]
      operationId: listImports
      summary: External locations, paged
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Import" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ArtifactMissing" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/exports:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/Query"
    get:
      tags: [results]
      operationId: listExports
      summary: External entry points, paged
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/Export" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ArtifactMissing" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/types:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/Query"
    get:
      tags: [results]
      operationId: listTypes
      summary: Data types, paged
      description: Structs, unions, enums, typedefs and function definitions.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/TypeEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ArtifactMissing" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/memory:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [results]
      operationId: listMemoryBlocks
      summary: Memory block index
      description: |
        Served whole, not paged. `bytes_exported` can be below `size`: the
        exporter stops at `GHIDRAREST_MAX_EXPORT_BYTES` across the program.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MemBlock" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/ArtifactMissing" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/hexdump/{addr}:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Addr"
    get:
      tags: [results]
      operationId: hexdump
      summary: Raw bytes at an address
      description: |
        Reads from the block dumps the export script wrote. Reads stop at the
        end of a block rather than running into the next one, which may not be
        contiguous.
      parameters:
        - name: length
          in: query
          description: Bytes to read. Clamped to `max_hexdump_bytes`.
          schema: { type: integer, minimum: 1, default: 256 }
        - name: format
          in: query
          schema:
            type: string
            enum: [json, text, raw]
            default: json
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Hexdump" }
            text/plain:
              schema:
                type: string
                description: "`hexdump -C` layout, program addresses in the left column."
            application/octet-stream:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Address is in no block, or the block is uninitialised (`.bss`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409": { $ref: "#/components/responses/NotDone" }
        "416":
          description: Inside a known block, but past its exported bytes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/results/{id}/signatures:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [signatures]
      operationId: signatures
      summary: Prototypes edited on this job
      description: |
        Also reports whether this job can be retyped at all, and which calling
        convention names its compiler spec defines.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignaturesResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/NotDone" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>`. Enforced only when
        `GHIDRAREST_API_TOKEN` is set; unset means the API is fully open.
        Comparison is constant time.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: The same shared token, for clients that prefer a custom header.
...
  responses:
...
    Unauthorized:
      description: A token is configured and the request did not carry it
      headers:
        WWW-Authenticate:
          schema: { type: string, examples: ['Bearer realm="ghidra-rest"'] }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such job
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Capabilities:
      type: object
      properties:
        service: { type: string }
        version: { type: string }
        ghidra_version: { type: string }
        limits:
          type: object
          properties:
            max_upload_bytes: { type: integer }
            max_concurrent: { type: integer }
            queue_size: { type: integer }
            analysis_timeout_sec: { type: integer }
            decompile_timeout_sec: { type: integer }
            decompile_max_funcs: { type: integer }
            max_export_bytes: { type: integer }
            max_page_size: { type: integer }
            max_hexdump_bytes: { type: integer }
        features: { type: array, items: { type: string } }
        endpoints: { type: array, items: { type: string } }
...